Move svg style methods into SVGStyle and svg utilities in SVGUtility
