+++ HTML_Renderer/src/main/java/org/lobobrowser/w3c/svg/SVGException.java
@@ -23,15 +23,18 @@ package org.lobobrowser.w3c.svg;
 public abstract class SVGException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;
+	
+	public short code;
+	
+	// ExceptionCode
+	public static final short SVG_WRONG_TYPE_ERR = 0;
+	
+	public static final short SVG_INVALID_VALUE_ERR = 1;
+	
+	public static final short SVG_MATRIX_NOT_INVERTABLE = 2;
 
 	public SVGException(short code, String message) {
 		super(message);
 		this.code = code;
 	}
-
-	public short code;
-	// ExceptionCode
-	public static final short SVG_WRONG_TYPE_ERR = 0;
-	public static final short SVG_INVALID_VALUE_ERR = 1;
-	public static final short SVG_MATRIX_NOT_INVERTABLE = 2;
 }
